Initial

+++ src/main/java/ir/afarinesh/realhope/shares/services/FileManagementService.java
@@ -10,18 +10,15 @@ import java.io.IOException;
 @Service
 public class FileManagementService {
 
-    public String pathSeparator(){
+    public String pathSeparator() {
         return "\\";
     }
 
-    public void createFile(String path, String fileName, String content) throws CreateFileException {
-        System.out.println("Path: " + path);
+    public void createFile(String path, String fileName, String content, boolean overwrite) throws CreateFileException {
         boolean isDirectoryAvailable;
         File directory = new File(path);
         if (!directory.exists()) {
-            System.out.println("Directory not exist ...");
             isDirectoryAvailable = directory.mkdirs();
-            System.out.println("isDirectoryAvailable = " + isDirectoryAvailable);
         } else {
             isDirectoryAvailable = true;
         }
@@ -29,15 +26,21 @@ public class FileManagementService {
         if (isDirectoryAvailable) {
             File file = new File(directory.getAbsolutePath() + this.pathSeparator() + fileName);
             try {
+                boolean exists = false;
                 if (file.exists()) {
-                    file.delete();
+                    exists = true;
+                    if (overwrite) {
+                        file.delete();
+                    }
                 }
-                if (file.createNewFile()) {
-                    FileWriter fileWriter = new FileWriter(file);
-                    fileWriter.write(content);
-                    fileWriter.close();
-                } else {
-                    throw new CreateFileException("The Create of the new file is failed");
+                if (!exists || overwrite) {
+                    if (file.createNewFile()) {
+                        FileWriter fileWriter = new FileWriter(file);
+                        fileWriter.write(content);
+                        fileWriter.close();
+                    } else {
+                        throw new CreateFileException("The Create of the new file is failed");
+                    }
                 }
             } catch (IOException e) {
                 throw new CreateFileException(e.getMessage());
